rename chapter number to 4
